# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/evaluations/list/all": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "课评"
                ],
                "summary": "课评列表[广场]",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "当前ID",
                        "name": "cur_evaluation_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "课评数量限制",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用于过滤课评的课程性质（可选）",
                        "name": "property",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/ginx.Result"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/evaluation.EvaluationVo"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/evaluations/list/courses/{courseId}": {
            "get": {
                "description": "根据课程ID获取课程评价列表，支持分页。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "课评"
                ],
                "summary": "课评列表[指定课程]",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "课程ID",
                        "name": "courseId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "当前课评ID",
                        "name": "cur_evaluation_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "返回课评的最大数量，上限为100",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/ginx.Result"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/evaluation.EvaluationVo"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/evaluations/list/mine": {
            "get": {
                "description": "根据课程ID获取课程评价列表，支持分页。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "课评"
                ],
                "summary": "课评列表[我的历史]",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "课程ID",
                        "name": "courseId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "当前评估ID，用于分页",
                        "name": "cur_evaluation_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "上限为100",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "课评状态: Public/Private/Folded",
                        "name": "status",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/ginx.Result"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/evaluation.EvaluationVo"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/evaluations/{evaluationId}/endorse": {
            "post": {
                "description": "根据评价ID支持或反对指定的评价",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "课评"
                ],
                "summary": "支持或反对课评",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "评价ID",
                        "name": "evaluationId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "支持或反对的请求体，stance为态度标识，-1反对，1支持，0表示无可用于取消表态",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/evaluation.EndorseReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "操作成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/ginx.Result"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "evaluation.EndorseReq": {
            "type": "object",
            "properties": {
                "stance": {
                    "type": "integer"
                }
            }
        },
        "evaluation.EvaluationVo": {
            "type": "object",
            "properties": {
                "assessments": {
                    "description": "考核方式，支持多选",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "content": {
                    "type": "string"
                },
                "course_id": {
                    "type": "integer"
                },
                "ctime": {
                    "type": "integer"
                },
                "features": {
                    "description": "课程特点，支持多选",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "publisher_id": {
                    "type": "integer"
                },
                "stance": {
                    "type": "integer"
                },
                "star_rating": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "total_comment_count": {
                    "type": "integer"
                },
                "total_oppose_count": {
                    "type": "integer"
                },
                "total_support_count": {
                    "type": "integer"
                },
                "utime": {
                    "type": "integer"
                }
            }
        },
...
        "ginx.Result": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "错误码，非 0 表示失败",
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "description": "错误或成功 描述",
                    "type": "string"
                }
            }
        },
...
    }
}